knowledge_extraction/requirements.txt
langchain
langchain-community
langchain-openai==0.1.17
openai==1.37.0
pgvector
psycopg==3.2.1
pypdf==4.3.1
python-dotenv==1.0.1
boto3
argparse
langchain_postgres

crewai
unstructured==0.10.25
pyowm==3.3.0
tools==0.1.9

gradio
